backend/services/ai-service/requirements.txt
# AI Service Requirements

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0

# ML (using sklearn instead of TensorFlow for Apple Silicon compatibility)
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.10.0
opencv-python-headless>=4.8.0

# OCR (EasyOCR uses PyTorch internally)
pytesseract>=0.3.10
easyocr>=1.7.0

# HTTP client for fetching images
httpx>=0.26.0
aiofiles>=23.2.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
